requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
future==1.0.0
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.31.0
requests-toolbelt==1.0.0
urllib3==2.2.1
webexteamssdk==1.6.1
